Reorganize buttons
